+++ week04_05/Leeyeeun/everytime/src/main/java/org/skhu/everytime/global/login/handler/LoginFailureHandler.java
@@ -8,10 +8,6 @@ import org.springframework.security.web.authentication.SimpleUrlAuthenticationFa
 
 import java.io.IOException;
 
-/**
- * JWT 로그인 실패 시 처리하는 핸들러
- * SimpleUrlAuthenticationFailureHandler를 상속받아서 구현
- */
 @Slf4j
 public class LoginFailureHandler extends SimpleUrlAuthenticationFailureHandler {
