+++ week04_05/Leeyeeun/everytime/src/main/java/org/skhu/everytime/global/oauth2/OAuthAttributes.java
@@ -2,9 +2,9 @@ package org.skhu.everytime.global.oauth2;
 
 import lombok.Builder;
 import lombok.Getter;
-import org.skhu.everytime.domain.user.Role;
-import org.skhu.everytime.domain.user.SocialType;
-import org.skhu.everytime.domain.user.User;
+import org.skhu.everytime.user.Role;
+import org.skhu.everytime.user.SocialType;
+import org.skhu.everytime.user.User;
 import org.skhu.everytime.global.oauth2.userinfo.GoogleOAuth2UserInfo;
 import org.skhu.everytime.global.oauth2.userinfo.KakaoOAuth2UserInfo;
 import org.skhu.everytime.global.oauth2.userinfo.OAuth2UserInfo;
@@ -12,10 +12,7 @@ import org.skhu.everytime.global.oauth2.userinfo.OAuth2UserInfo;
 import java.util.Map;
 import java.util.UUID;
 
-/**
- * 각 소셜에서 받아오는 데이터가 다르므로
- * 소셜별로 데이터를 받는 데이터를 분기 처리하는 DTO 클래스
- */
+
 @Getter
 public class OAuthAttributes {
 
@@ -28,12 +25,7 @@ public class OAuthAttributes {
         this.oauth2UserInfo = oauth2UserInfo;
     }
 
-    /**
-     * SocialType에 맞는 메소드 호출하여 OAuthAttributes 객체 반환
-     * 파라미터 : userNameAttributeName -> OAuth2 로그인 시 키(PK)가 되는 값 / attributes : OAuth 서비스의 유저 정보들
-     * 소셜별 of 메소드(ofGoogle, ofKaKao, ofNaver)들은 각각 소셜 로그인 API에서 제공하는
-     * 회원의 식별값(id), attributes, nameAttributeKey를 저장 후 build
-     */
+
     public static OAuthAttributes of(SocialType socialType,
                                      String userNameAttributeName, Map<String, Object> attributes) {
 
@@ -58,12 +50,6 @@ public class OAuthAttributes {
     }
 
 
-    /**
-     * of메소드로 OAuthAttributes 객체가 생성되어, 유저 정보들이 담긴 OAuth2UserInfo가 소셜 타입별로 주입된 상태
-     * OAuth2UserInfo에서 socialId(식별값), nickname, imageUrl을 가져와서 build
-     * email에는 UUID로 중복 없는 랜덤 값 생성
-     * role은 GUEST로 설정
-     */
     public User toEntity(SocialType socialType, OAuth2UserInfo oauth2UserInfo) {
         return User.builder()
                 .socialType(socialType)
