+++ week04_05/Leeyeeun/everytime/src/main/java/org/skhu/everytime/global/oauth2/handler/OAuth2LoginSuccessHandler.java
@@ -6,7 +6,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.skhu.everytime.domain.user.Role;
+import org.skhu.everytime.user.Role;
 import org.skhu.everytime.global.jwt.service.JwtService;
 import org.skhu.everytime.global.oauth2.CustomOAuth2User;
 import org.springframework.security.core.Authentication;
@@ -50,7 +50,6 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
 
     }
 
-    // TODO : 소셜 로그인 시에도 무조건 토큰 생성하지 말고 JWT 인증 필터처럼 RefreshToken 유/무에 따라 다르게 처리해보기
     private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
         String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
         String refreshToken = jwtService.createRefreshToken();
