+++ week04_05/Leeyeeun/everytime/src/main/java/org/skhu/everytime/global/oauth2/userinfo/OAuth2UserInfo.java
@@ -10,7 +10,7 @@ public abstract class OAuth2UserInfo {
         this.attributes = attributes;
     }
 
-    public abstract String getId(); //소셜 식별 값 : 구글 - "sub", 카카오 - "id", 네이버 - "id"
+    public abstract String getId();
 
     public abstract String getNickname();
